Java 21 and other improvements
- Allow reading unencrypted *.pak files in zip format, which the client supports natively (basically adds support for modded clients)
- Dependency updates
- Maven config to build an executable jar file
- Fixed text render issues with some terminals
- Refactoring

+++ src/main/java/com/aionemu/geobuilder/entries/HouseEntry.java
@@ -6,7 +6,7 @@ import java.util.stream.Stream;
 
 public class HouseEntry extends EntityEntry {
 
-  public int address;
+  public short address;
   public List<String> meshes = new ArrayList<>();
 
   public HouseEntry() {
